+++ src/main/java/com/sloydev/sevibus/api/data/arrivals/AppTussamArrivalTimesRepository.java
@@ -1,12 +1,13 @@
-package com.sloydev.sevibus.api.data.apptusam;
+package com.sloydev.sevibus.api.data.arrivals;
 
 
-import com.sloydev.sevibus.api.data.apptusam.model.Envelope;
-import com.sloydev.sevibus.api.data.apptusam.model.Estimacion;
-import com.sloydev.sevibus.api.data.apptusam.model.TiempoLinea;
-import com.sloydev.sevibus.api.domain.ArrivalTimes;
-import com.sloydev.sevibus.api.domain.ArrivalTimesException;
-import com.sloydev.sevibus.api.domain.ArrivalTimesRepository;
+import com.sloydev.sevibus.api.data.internal.apptusam.AppTussamApi;
+import com.sloydev.sevibus.api.data.internal.apptusam.arrivals.model.Envelope;
+import com.sloydev.sevibus.api.data.internal.apptusam.arrivals.model.Estimacion;
+import com.sloydev.sevibus.api.data.internal.apptusam.arrivals.model.TiempoLinea;
+import com.sloydev.sevibus.api.domain.arrivals.ArrivalTimes;
+import com.sloydev.sevibus.api.domain.arrivals.ArrivalTimesException;
+import com.sloydev.sevibus.api.domain.arrivals.ArrivalTimesRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
