+++ src/main/java/com/sloydev/sevibus/api/domain/arrivals/ArrivalTimesException.java
@@ -1,4 +1,4 @@
-package com.sloydev.sevibus.api.domain;
+package com.sloydev.sevibus.api.domain.arrivals;
 
 
 public class ArrivalTimesException extends RuntimeException{
