+++ src/main/java/com/sloydev/sevibus/api/domain/arrivals/ArrivalTimesRepository.java
@@ -1,4 +1,4 @@
-package com.sloydev.sevibus.api.domain;
+package com.sloydev.sevibus.api.domain.arrivals;
 
 
 import java.util.List;
